+++ Modu.lua
@@ -17,7 +17,7 @@ local owner = nil -- Your MC name
 ---------------------START: Initialization---------------------
 
 local modules = {}
-local initData = require("modules.InitData")
+local initData = require("data.InitData")
 modules.Inventory = require("modules.Inventory")
 modules.Chat = require("modules.Chat")
